# docs/api-spec/openapi.yml
openapi: 3.0.3
info:
  title: Authorization API
  description: 認可サーバーのAPI群です。
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: API開発チーム
    email: satoshi.nagashiba@gmail.com
servers:
  - url: https://apis.authorization.com
    description: 本番環境
  - url: https://apis-stg.authorization.com
    description: ステージング環境
  - url: https://apis-dev.authorization.com
    description: 開発環境
  - url: https://apis.authorization-php.dev/api
    description: ローカルPHP（Laravel）環境
  - url: https://apis.authorization-csharp.dev/api
    description: ローカルC#（ASP.NET Core）環境
  - url: https://apis.authorization-go-gin.dev/api
    description: ローカルGo（Gin）環境
  - url: https://apis.authorization-go-beego.dev/api
    description: ローカルGo（Beego）環境
  - url: https://apis.authorization-go-echo.dev/api
    description: ローカルGo（Echo）環境
  - url: https://apis.authorization-java.dev/api
    description: ローカルJava（Spring Boot）環境
  - url: https://apis.authorization-kotlin.dev/api
    description: ローカルKotlin（Ktor）環境
  - url: https://apis.authorization-python.dev/api
    description: ローカルPython（FastAPI）環境
  - url: https://apis.authorization-rb-hanami.dev/api
    description: ローカルRuby（Hanami）環境
  - url: https://apis.authorization-rb-rails.dev/api
    description: ローカルRuby（Rails）環境
  - url: https://apis.authorization-rust.dev/api
    description: ローカルRust（Axum）環境
  - url: https://apis.authorization-ts.dev/api
    description: ローカルTypeScript（Hono）環境
paths:
  /auth/login:
    get:
      operationId: getLogin
      summary: ログイン情報を取得します
      tags:
        - auth
      security: []
      responses:
        '200':
          description: 成功時にレスポンスを返します。
        '404':
          description: ログインページが存在しない場合のエラー
  /auth/invitation/{token}:
    get:
      operationId: getInvitation
      summary: 招待URLトークンを検証し、ログイン情報を取得します
      tags:
        - auth
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: 招待トークン
      responses:
        '200':
          description: 成功時にレスポンスを返します。
        '404':
          description: ログインページが存在しない場合のエラー
        '500':
          description: 招待トークンが不正な場合のエラー
  /auth/google/redirect:
    servers:
      - url: https://apis.authorization.com
        description: 本番環境
      - url: https://apis-stg.authorization.com
        description: ステージング環境
      - url: https://apis-dev.authorization.com
        description: 開発環境
      - url: https://apis.authorization-php.dev
        description: ローカルPHP（Laravel）環境
      - url: https://apis.authorization-csharp.dev
        description: ローカルC#（ASP.NET Core）環境
      - url: https://apis.authorization-go-gin.dev
        description: ローカルGo（Gin）環境
      - url: https://apis.authorization-go-beego.dev
        description: ローカルGo（Beego）環境
      - url: https://apis.authorization-go-echo.dev
        description: ローカルGo（Echo）環境
      - url: https://apis.authorization-java.dev
        description: ローカルJava（Spring Boot）環境
      - url: https://apis.authorization-kotlin.dev
        description: ローカルKotlin（Ktor）環境
      - url: https://apis.authorization-python.dev
        description: ローカルPython（FastAPI）環境
      - url: https://apis.authorization-rb-hanami.dev
        description: ローカルRuby（Hanami）環境
      - url: https://apis.authorization-rb-rails.dev
        description: ローカルRuby（Rails）環境
      - url: https://apis.authorization-rust.dev
        description: ローカルRust（Axum）環境
      - url: https://apis.authorization-ts.dev
        description: ローカルTypeScript（Hono）環境
    get:
      operationId: googleRedirect
      summary: Googleログインページへリダイレクトします
      description: |
        ブラウザリダイレクトのため /api 外に配置。
        CSRF 対策として暗号学的に安全な nonce を生成し、HttpOnly クッキー `oauth_state`（有効期限 10 分）に保存したうえで
        state パラメータ（`{runtime}|{nonce}` または招待時 `{runtime}|{nonce}|{token}`）に埋め込む。
      tags:
        - auth
      security: []
      responses:
        '302':
          description: Google 認証ページへリダイレクト
        '400':
          description: リクエストが不正です
        '500':
          description: Googleログインに失敗した場合のエラー
  /auth/google/callback:
    servers:
      - url: https://apis.authorization.com
        description: 本番環境
      - url: https://apis-stg.authorization.com
        description: ステージング環境
      - url: https://apis-dev.authorization.com
        description: 開発環境
      - url: https://apis.authorization-php.dev
        description: ローカルPHP（Laravel）環境
      - url: https://apis.authorization-csharp.dev
        description: ローカルC#（ASP.NET Core）環境
      - url: https://apis.authorization-go-gin.dev
        description: ローカルGo（Gin）環境
      - url: https://apis.authorization-go-beego.dev
        description: ローカルGo（Beego）環境
      - url: https://apis.authorization-go-echo.dev
        description: ローカルGo（Echo）環境
      - url: https://apis.authorization-java.dev
        description: ローカルJava（Spring Boot）環境
      - url: https://apis.authorization-kotlin.dev
        description: ローカルKotlin（Ktor）環境
      - url: https://apis.authorization-python.dev
        description: ローカルPython（FastAPI）環境
      - url: https://apis.authorization-rb-hanami.dev
        description: ローカルRuby（Hanami）環境
      - url: https://apis.authorization-rb-rails.dev
        description: ローカルRuby（Rails）環境
      - url: https://apis.authorization-rust.dev
        description: ローカルRust（Axum）環境
      - url: https://apis.authorization-ts.dev
        description: ローカルTypeScript（Hono）環境
    get:
      operationId: googleCallback
      summary: Googleログインからのコールバックを受け取ります
      description: |
        ブラウザリダイレクトのため /api 外に配置。
        state に含まれる nonce と `oauth_state` クッキーの値を定数時間比較で検証し、
        クッキー欠落・state 不正・nonce 不一致の場合は /error?code=400 へリダイレクトする。
        検証後は成否を問わず `oauth_state` クッキーを破棄する。
      tags:
        - auth
      security: []
      responses:
        '302':
          description: ログイン成功後フロントエンドへリダイレクト（state 検証失敗時は /error?code=400 へリダイレクト）
        '403':
          description: 招待フローで招待トークンが無効な場合
        '500':
          description: Googleログインに失敗した場合のエラー
  /auth/github/redirect:
    servers:
      - url: https://apis.authorization.com
        description: 本番環境
      - url: https://apis-stg.authorization.com
        description: ステージング環境
      - url: https://apis-dev.authorization.com
        description: 開発環境
      - url: https://apis.authorization-php.dev
        description: ローカルPHP（Laravel）環境
      - url: https://apis.authorization-csharp.dev
        description: ローカルC#（ASP.NET Core）環境
      - url: https://apis.authorization-go-gin.dev
        description: ローカルGo（Gin）環境
      - url: https://apis.authorization-go-beego.dev
        description: ローカルGo（Beego）環境
      - url: https://apis.authorization-go-echo.dev
        description: ローカルGo（Echo）環境
      - url: https://apis.authorization-java.dev
        description: ローカルJava（Spring Boot）環境
      - url: https://apis.authorization-kotlin.dev
        description: ローカルKotlin（Ktor）環境
      - url: https://apis.authorization-python.dev
        description: ローカルPython（FastAPI）環境
      - url: https://apis.authorization-rb-hanami.dev
        description: ローカルRuby（Hanami）環境
      - url: https://apis.authorization-rb-rails.dev
        description: ローカルRuby（Rails）環境
      - url: https://apis.authorization-rust.dev
        description: ローカルRust（Axum）環境
      - url: https://apis.authorization-ts.dev
        description: ローカルTypeScript（Hono）環境
    get:
      operationId: githubRedirect
      summary: GitHubログインページへリダイレクトします
      description: |
        ブラウザリダイレクトのため /api 外に配置。
        state パラメータにバックエンド識別子（例: go-gin）を含め、
        GitHub からのコールバックを正しいバックエンドへ振り分けるために使用する。
        あわせて CSRF 対策として暗号学的に安全な nonce を生成し、HttpOnly クッキー `oauth_state`（有効期限 10 分）に保存したうえで
        state（`{runtime}|{nonce}` または招待時 `{runtime}|{nonce}|{token}`）に埋め込む。
      tags:
        - auth
      security: []
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
          description: 招待トークン（招待フロー経由の場合のみ）
      responses:
        '302':
          description: GitHub 認証ページへリダイレクト
        '400':
          description: リクエストが不正です
        '500':
          description: GitHubログインに失敗した場合のエラー
  /auth/github/callback:
    servers:
      - url: https://apis.authorization.com
        description: 本番環境
      - url: https://apis-stg.authorization.com
        description: ステージング環境
      - url: https://apis-dev.authorization.com
        description: 開発環境
      - url: https://apis.authorization-php.dev
        description: ローカルPHP（Laravel）環境
      - url: https://apis.authorization-csharp.dev
        description: ローカルC#（ASP.NET Core）環境
      - url: https://apis.authorization-go-gin.dev
        description: ローカルGo（Gin）環境
      - url: https://apis.authorization-go-beego.dev
        description: ローカルGo（Beego）環境
      - url: https://apis.authorization-go-echo.dev
        description: ローカルGo（Echo）環境
      - url: https://apis.authorization-java.dev
        description: ローカルJava（Spring Boot）環境
      - url: https://apis.authorization-kotlin.dev
        description: ローカルKotlin（Ktor）環境
      - url: https://apis.authorization-python.dev
        description: ローカルPython（FastAPI）環境
      - url: https://apis.authorization-rb-hanami.dev
        description: ローカルRuby（Hanami）環境
      - url: https://apis.authorization-rb-rails.dev
        description: ローカルRuby（Rails）環境
      - url: https://apis.authorization-rust.dev
        description: ローカルRust（Axum）環境
      - url: https://apis.authorization-ts.dev
        description: ローカルTypeScript（Hono）環境
    get:
      operationId: githubCallback
      summary: GitHubログインからのコールバックを受け取ります
      description: |
        ブラウザリダイレクトのため /api 外に配置。
        GitHub は redirect_uri を1つしか登録できないため、フロントエンドの
        /auth/github/callback（Next.js ディスパッチャー）が state を読み取り、
        /function/{backend}/auth/github/callback へ転送してからこのエンドポイントに到達する。
        state に含まれる nonce と `oauth_state` クッキーの値を定数時間比較で検証し、
        クッキー欠落・state 不正・nonce 不一致の場合は /error?code=400 へリダイレクトする。
        検証後は成否を問わず `oauth_state` クッキーを破棄する。
      tags:
        - auth
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: GitHub から発行される認可コード
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: "`{runtime}|{nonce}` または `{runtime}|{nonce}|{token}`。先頭のバックエンド識別子（例: go-gin）をディスパッチャーがルーティングに使用し、nonce を CSRF 検証に使用する"
      responses:
        '302':
          description: ログイン成功後フロントエンドへリダイレクト（state 検証失敗時は /error?code=400 へリダイレクト）
        '403':
          description: 招待フローで招待トークンが無効な場合
        '500':
          description: GitHubログインに失敗した場合のエラー
  /auth/me:
    get:
      operationId: getMe
      summary: 自分自身のプロフィールを取得します
      description: staff_id クッキーで認証済みのユーザー情報を返します。ログイン判定・ヘッダー表示などに使用します。
      tags:
        - auth
      security: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMe'
        '401':
          description: 未認証（staff_id クッキーなし、または復号失敗）
        '404':
          description: スタッフが存在しない
  /auth/logout:
    get:
      operationId: logout
      summary: ログアウトします
      tags:
        - auth
      security: []
      responses:
        '200':
          description: 成功時にレスポンスを返します。
        '401':
          description: 未認証
        '500':
          description: ログアウトに失敗した場合のエラー
...
components:
  schemas:
    AuthMe:
      type: object
      description: ログイン中スタッフのプロフィール情報
      required:
        - staff_id
        - name
      properties:
        staff_id:
          type: integer
          format: int64
          description: スタッフID
          example: 1
        name:
          type: string
          description: 表示名
          example: "長柴聡史"
        avatar:
          type: string
          nullable: true
          description: アバター画像URL（未設定時は null）
          example: "https://lh3.googleusercontent.com/a/example"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    clientTokenAuth:
      type: http
      scheme: bearer
      description: クライアントのアクセストークン認証（JWT ではなく登録時に発行される固定トークン）
...